restructure everything
