+++ main.py
@@ -1,8 +1,8 @@
-import get_data
-import student_info
-import string_functions
-import handle_query
-import find_DB
+from attendance import get_data
+from attendance import student_info
+from attendance import string_functions
+from attendance import handle_query
+from db import find_DB
 
 try:
     user = student_info.Student()
